requirements.txt
torch>=1.12.0
numpy
pyyaml
tqdm
pytest
matplotlib
torchvision>=0.13.0
